Move tests into a package
