+++ data-prepper-plugins/mutate-event/src/test/java/com/amazon/dataprepper/plugins/processor/mutateevent/AddProcessorTests.java
@@ -3,12 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+package com.amazon.dataprepper.plugins.processor.mutateevent;
+
 import com.amazon.dataprepper.metrics.PluginMetrics;
 import com.amazon.dataprepper.model.event.Event;
 import com.amazon.dataprepper.model.event.JacksonEvent;
 import com.amazon.dataprepper.model.record.Record;
-import com.amazon.dataprepper.plugins.processor.mutateevent.CopyProcessor;
-import com.amazon.dataprepper.plugins.processor.mutateevent.CopyProcessorConfig;
 import org.junit.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.runner.RunWith;
@@ -29,42 +29,38 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 @RunWith(MockitoJUnitRunner.class)
-public class CopyProcessorTests {
+public class AddProcessorTests {
     @Mock
     private PluginMetrics pluginMetrics;
 
     @Mock
-    private CopyProcessorConfig mockConfig;
+    private AddProcessorConfig mockConfig;
 
     @InjectMocks
-    private CopyProcessor processor;
+    private AddProcessor processor;
 
     @Test
-    public void testCopyProcessorTests() {
-        when(mockConfig.getFrom()).thenReturn("message");
-        when(mockConfig.getTo()).thenReturn("newMessage");
+    public void testAddProcessorTests() {
+        when(mockConfig.getKey()).thenReturn("newMessage");
+        when(mockConfig.getValue()).thenReturn(3);
         final Record<Event> record = getMessage("thisisamessage");
         final List<Record<Event>> editedRecords = (List<Record<Event>>) processor.doExecute(Collections.singletonList(record));
 
         assertThat(editedRecords.get(0).getData().containsKey("newMessage"), is(true));
-        assertThat(editedRecords.get(0).getData().containsKey("message"), is(true));
-        assertThat(editedRecords.get(0).getData().get("newMessage", Object.class), equalTo("thisisamessage"));
-        assertThat(editedRecords.get(0).getData().get("message", Object.class), equalTo("thisisamessage"));
+        assertThat(editedRecords.get(0).getData().get("newMessage", Object.class), equalTo(3));
     }
 
     @Test
-    public void testCopyNoOverwriteProcessorTests() {
-        when(mockConfig.getFrom()).thenReturn("message");
-        when(mockConfig.getTo()).thenReturn("newMessage");
+    public void testAddNoOverwriteProcessorTests() {
+        when(mockConfig.getKey()).thenReturn("newMessage");
+        when(mockConfig.getValue()).thenReturn(3);
         when(mockConfig.getSkipIfPresent()).thenReturn(true);
         final Record<Event> record = getMessage("thisisamessage");
         record.getData().put("newMessage", "test");
         final List<Record<Event>> editedRecords = (List<Record<Event>>) processor.doExecute(Collections.singletonList(record));
 
         assertThat(editedRecords.get(0).getData().containsKey("newMessage"), is(true));
-        assertThat(editedRecords.get(0).getData().containsKey("message"), is(true));
         assertThat(editedRecords.get(0).getData().get("newMessage", Object.class), equalTo("test"));
-        assertThat(editedRecords.get(0).getData().get("message", Object.class), equalTo("thisisamessage"));
     }
 
     private Record<Event> getMessage(String message) {
